move ServerRequest to hooks
