+++ packages/kit/types/index.d.ts
@@ -4,6 +4,12 @@
 import './ambient-modules';
 
 export { Adapter, AdapterUtils, Config } from './config';
+export { EndpointOutput, RequestHandler } from './endpoint';
 export { ErrorLoad, Load, Page } from './page';
-export { Incoming, GetSession, Handle, ServerResponse as Response } from './hooks';
-export { ServerRequest as Request, EndpointOutput, RequestHandler } from './endpoint';
+export {
+	Incoming,
+	GetSession,
+	Handle,
+	ServerRequest as Request,
+	ServerResponse as Response
+} from './hooks';
